+++ contracts/accounts/interchain_account/Cargo.toml
@@ -24,5 +24,5 @@ serde                 = { workspace = true }
 thiserror             = { workspace = true }
 neutron-sdk           = { workspace = true }
 serde_json            = { workspace = true }
+valence-account-utils = { workspace = true }
 valence-ibc-utils     = { workspace = true, features = ["neutron"] }
-valence-account-utils = { workspace = true, features = ["neutron"] }
