+++ contracts/testing/test-icq-lib/src/state.rs
@@ -1,5 +1,5 @@
-use cosmwasm_schema::cw_serde;
 use cw_storage_plus::{Item, Map};
+use valence_ibc_utils::neutron::Transfer;
 
 /// contains all transfers mapped by a recipient address observed by the contract.
 pub const RECIPIENT_TXS: Map<String, Vec<Transfer>> = Map::new("recipient_txs");
@@ -7,21 +7,3 @@ pub const RECIPIENT_TXS: Map<String, Vec<Transfer>> = Map::new("recipient_txs");
 pub const TRANSFERS: Item<u64> = Item::new("transfers");
 
 pub const CATCHALL: Map<String, String> = Map::new("catchall");
-
-#[cw_serde]
-pub struct Transfer {
-    pub recipient: String,
-    pub sender: String,
-    pub denom: String,
-    pub amount: String,
-}
-
-#[cw_serde]
-pub struct OpenAckVersion {
-    pub version: String,
-    pub controller_connection_id: String,
-    pub host_connection_id: String,
-    pub address: String,
-    pub encoding: String,
-    pub tx_type: String,
-}
