+++ packages/account-utils/src/lib.rs
@@ -1,8 +1,5 @@
 pub mod error;
 pub mod msg;
 
-#[cfg(feature = "neutron")]
-pub mod ica;
-
 #[cfg(feature = "testing")]
 pub mod testing;
